+++ projects/bela.PAM/dsp/pam-rotary-effect.dsp
@@ -1,0 +1,5 @@
+declare author "Jean-Loup Pecquais";
+declare copyright "MyCompany";
+declare version "1.00";
+
+process = component("../../faust/modules/modulations/rotary-speaker.dsp");
