--- projects/bela.PAM/model training/.gitignore
@@ -1,3 +1,0 @@
-TensorboardData
-Data
-Results
