--- projects/bela.PAM/model training/CoreAudioML/README.md
@@ -1,7 +1,0 @@
-# CoreAudioML
-
-![Module Tests](https://github.com/Alec-Wright/CoreAudioML/workflows/Python%20Package%20using%20Conda/badge.svg?branch=main)
-
-### Machine learning for audio effects processing
-
-This repo contains functions and classes used in training neural network models of audio effects such as guitar amplifiers, distortion effects, phasers and flangers. This repo is intended to be used as a submodule for the other repos that actually contain the trained models and training scripts, for example [amplifiers/distortion](https://github.com/Alec-Wright/NeuralGuitarAmpModelling) and [phaser/flanger](https://github.com/Alec-Wright/NeuralTimeVaryFX)
