--- projects/bela.PAM/model training/CoreAudioML/result_test/network1/TestConfig.json
@@ -1,1 +1,0 @@
-{"unit_type": "LSTM", "hidden_size": 32, "in_size": 2, "out_size": 1, "num_layers": 1}
