--- projects/bela.PAM/model training/CoreAudioML/result_test/network1/tloss.txt
@@ -1,1 +1,0 @@
-[0.01873958]
