+++ projects/bela.PAM/training/.gitignore
@@ -1,0 +1,3 @@
+TensorboardData
+Data
+Results
