+++ projects/bela.PAM/training/Configs/RNN3-bela.json
@@ -1,0 +1,1 @@
+{"hidden_size": 8, "unit_type": "GRU", "loss_fcns": {"ESR": 0.75, "DC": 0.25}, "pre_filt": "None", "device": "bela", "file_name": "bela"}
