+++ projects/bela.PAM/training/requirements.txt
@@ -1,0 +1,4 @@
+numpy==1.23.5
+scipy==1.14.1
+soundfile==0.12.1
+torch==1.11.0
